requirements.txt
streamlit>=1.36
langgraph>=0.2.60
langchain-core>=0.3.15
replicate>=1.0
beautifulsoup4>=4.12
lxml>=5.0
sentence-transformers>=3.0
psycopg[binary,pool]>=3.2
pgvector>=0.3.6
pydantic>=2.8
pydantic-settings>=2.4
python-dotenv>=1.0
pypdf>=5.0
